Rename createGlobalLookupDictionary.ts and test
