+++ src/utils/transformingDictionaries/createGlobalLookupDictionary.test.ts
@@ -1,4 +1,4 @@
-import { createGlobalLookupDictionary } from "./createAGlobalLookupDictionary";
+import { createGlobalLookupDictionary } from "./createGlobalLookupDictionary";
 import { AffixList } from "../affixList";
 import {
   testTypeyTypeDict,
