+++ src/utils/transformingDictionaries/createGlobalLookupDictionary.ts
