+++ src/utils/transformingDictionaries/recursiveBuildStrokeHint.test.ts
@@ -1,5 +1,5 @@
 import recursiveBuildStrokeHint from "./recursiveBuildStrokeHint";
-import { createGlobalLookupDictionary } from "./createAGlobalLookupDictionary";
+import { createGlobalLookupDictionary } from "./createGlobalLookupDictionary";
 import { AffixList } from "../affixList";
 import {
   testTypeyTypeDict,
