+++ src/utils/transformingDictionaries/transformingDictionaries.test.ts
@@ -2,7 +2,7 @@ import {
   addOutlinesToWordsInCombinedDict,
   generateListOfWordsAndStrokes,
 } from "./transformingDictionaries";
-import { createGlobalLookupDictionary } from "./createAGlobalLookupDictionary";
+import { createGlobalLookupDictionary } from "./createGlobalLookupDictionary";
 import { AffixList } from "../affixList";
 import {
   testTypeyTypeDict,
